fix: test 미통과 버그 수정 및 react-query setting 변경
