--- src/app/data.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-import { queryOptions } from "@tanstack/react-query";
-
-const getData = queryOptions({
-  queryKey: ["posts"],
-  queryFn: async () => {
-    const response = await fetch("경로");
-
-    return response.json();
-  },
-});
-
-export default getData;
