+++ src/app/page.tsx
@@ -1,22 +1,5 @@
-import {
-  dehydrate,
-  HydrationBoundary,
-  QueryClient,
-} from "@tanstack/react-query";
-// import getData from "./data";
-
 const Home = () => {
-  const queryClient = new QueryClient();
-  // queryClient.prefetchQuery(getData);
-
-  return (
-    <main>
-      메인 페이지 초기화
-      <HydrationBoundary state={dehydrate(queryClient)}>
-        {/** 컴포넌트 예정 */}
-      </HydrationBoundary>
-    </main>
-  );
+  return <main>메인 페이지 초기화</main>;
 };
 
 export default Home;
