--- src/components/app/getQueryClient.ts
@@ -1,34 +1,0 @@
-import {
-  QueryClient,
-  defaultShouldDehydrateQuery,
-  isServer,
-} from "@tanstack/react-query";
-
-const DEFAULT_STALE_TIME = 60 * 1000;
-
-function makeQueryClient() {
-  return new QueryClient({
-    defaultOptions: {
-      queries: {
-        staleTime: DEFAULT_STALE_TIME,
-      },
-      dehydrate: {
-        shouldDehydrateQuery: (query) =>
-          defaultShouldDehydrateQuery(query) ||
-          query.state.status === "pending",
-      },
-    },
-  });
-}
-
-let browserQueryClient: QueryClient | undefined;
-
-const getQueryClient = () => {
-  if (isServer) {
-    return makeQueryClient();
-  }
-  if (!browserQueryClient) browserQueryClient = makeQueryClient();
-  return browserQueryClient;
-};
-
-export default getQueryClient;
